*refactoring of ConvertLayerNormalizationNode.
