+++ prj/vs2022/CvtOnnx.vcxproj.filters
@@ -159,6 +159,9 @@
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertConcatNode.cpp">
       <Filter>OnnxRuntime\C</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertLayerNormalizationNode.cpp">
+      <Filter>OnnxRuntime\L</Filter>
+    </ClCompile>
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertLessNode.cpp">
       <Filter>OnnxRuntime\L</Filter>
     </ClCompile>
